+++ shotmanager/features/greasepencil/__init__.py
@@ -26,6 +26,8 @@ from . import greasepencil_properties
 from . import greasepencil_operators
 from . import greasepencil_tools
 
+# from . import greasepencil_toolboxTest_ui
+
 from shotmanager.config import sm_logging
 
 _logger = sm_logging.getLogger(__name__)
@@ -43,7 +45,10 @@ def register():
     greasepencil_properties.register()
     greasepencil_operators.register()
     greasepencil_tools.register()
-    # rendering_ui.register()    # done in shotmanager.__init__ in order to display the panel in the right order
+
+
+#  greasepencil_toolbox_ui.register()
+# rendering_ui.register()    # done in shotmanager.__init__ in order to display the panel in the right order
 
 
 def unregister():
@@ -53,6 +58,7 @@ def unregister():
     #     bpy.utils.unregister_class(cls)
 
     # rendering_ui.unregister()   # done in shotmanager.__init__ in order to display the panel in the right order
+    # greasepencil_toolbox_ui.unregister()
     greasepencil_tools.unregister()
     greasepencil_operators.unregister()
     greasepencil_properties.unregister()