+++ shotmanager/features/greasepencil/greasepencil_ui.py
@@ -23,6 +23,8 @@ from shotmanager.utils import utils
 from shotmanager.utils import utils_ui
 from shotmanager.utils import utils_greasepencil
 
+from shotmanager.features.greasepencil import greasepencil_toolspanel_ui as gp
+
 from shotmanager.config import config
 
 
@@ -57,11 +59,17 @@ def draw_greasepencil_shot_properties(layout, context, shot):
     if len(shot.greasePencils):
         gpProperties = shot.greasePencils[0]
 
+    gp_child = utils_greasepencil.get_greasepencil_child(shot.camera)
+
     box = layout.box()
     box.use_property_decorate = False
     row = box.row(align=True)
 
-    gp_child = utils_greasepencil.get_greasepencil_child(shot.camera)
+    # grease pencil
+    ################
+
+    if props.display_greasepencil_in_properties:  # and props.expand_greasepencil_properties:
+        gp.draw_greasepencil_play_tools(layout, context, shot, layersListDropdown=prefs.layersListDropdown)
 
     extendSubRow = row.row(align=True)
     # extendSubRow.alignment = "RIGHT"
@@ -95,29 +103,28 @@ def draw_greasepencil_shot_properties(layout, context, shot):
     else:
         extendSubRow.alignment = "EXPAND"
         subRow = extendSubRow.row(align=False)
-        # subRow.scale_x = 0.8
-        selSubRow = subRow.row(align=True)
-        selSubRow.label(text=gp_child.name)
-
-        # Visibility tools
-        ################
-        subSubRow = subRow.row(align=True)
-        subSubRow.prop(gp_child, "hide_select", text="")
-        subSubRow.prop(gp_child, "hide_viewport", text="")
-        subSubRow.prop(gp_child, "hide_render", text="")
-
-        subRow = row.row(align=True)
-        # subRow.alignment = "RIGHT"
         subRow.separator(factor=0.9)
         subRow.operator("uas_shot_manager.remove_grease_pencil", text="", icon="PANEL_CLOSE").shotIndex = shotIndex
-        subRow.separator()
         subRow.prop(props, "display_greasepencil_in_shotlist", text="")
+
+        # name and visibility tools
+        ################
+        # extendSubRow.alignment = "EXPAND"
+        subRow = box.row(align=False)
+        # subRow.scale_x = 0.8
+        leftSubRow = subRow.row(align=True)
+        leftSubRow.alignment = "LEFT"
+        leftSubRow.label(text="GP Object: ")
+        leftSubRow.label(text=gp_child.name)
+
+        rightSubRow = subRow.row(align=True)
+        rightSubRow.alignment = "RIGHT"
+        rightSubRow.prop(gp_child, "hide_select", text="")
+        rightSubRow.prop(gp_child, "hide_viewport", text="")
+        rightSubRow.prop(gp_child, "hide_render", text="")
 
         # Tools
         ################
-        mainRow = extendSubRow.row()
-        hSepRow = mainRow.row()
-        hSepRow.separator(factor=0.5)
 
         mainRow = box.row()
         mainRow.separator(factor=leftSepFactor)
@@ -182,8 +189,19 @@ def draw_greasepencil_shot_properties(layout, context, shot):
             transformRow.prop(gp_child, "scale")
 
             transformRow = col.row()
-            transformRow.operator("object.paths_update", text="Update Paths", icon="OBJECT_DATA")
-            transformRow.operator("object.paths_clear", text="", icon="X")
+            transformRow.separator(factor=0.6)
+            transformRow = col.row(align=True)
+            transformRow.alignment = "RIGHT"
+            # transformRow.ui_units_x = 2
+            if gp_child.motion_path is None:
+                transformRow.operator("object.paths_calculate", text="Calculate...", icon="OBJECT_DATA")
+            else:
+                # transformRow.operator("object.paths_update", text="Update Paths", icon="OBJECT_DATA")
+                transformRow.operator("object.paths_update_visible", text="Update All Paths", icon="WORLD")
+                transformRow.operator("object.paths_clear", text="", icon="X")
+
+            row = col.row()
+            row.separator(factor=0.4)
 
         row = col.row()
         row.separator(factor=hSepFactor)
@@ -240,156 +258,3 @@ def draw_greasepencil_global_properties(layout, context):
     c.operator("uas_shot_manager.remove_grease_pencil", text="", icon="PANEL_CLOSE")
 
     col.separator(factor=0.5)
-
-
-def draw_greasepencil_play_tools(layout, context, shot, layersListDropdown=None):
-    props = context.scene.UAS_shot_manager_props
-    # prefs = context.preferences.addons["shotmanager"].preferences
-
-    if shot is None:
-        return
-
-    utils_ui.drawSeparatorLine(layout)
-
-    box = layout.box()
-    #    row = box.row()
-
-    # Object and layers ###
-    ###################
-
-    mainRow = box.row(align=False)
-
-    objIsGP = False
-    selObjName = "-"
-    if context.object is not None:
-        selObjName = context.object.name
-        objIsGP = "GPENCIL" == context.object.type
-        if objIsGP:
-            gp = context.object
-    # mainRow.label(text=f"Sel: {selObjName}, GP: {objIsGP}")
-    mainRow.label(text=f"GP:  {gp.name if objIsGP else '-'}")
-
-    icon = "GP_SELECT_STROKES"
-
-    if objIsGP:
-        layersRow = mainRow.row(align=True)
-        layersRow.alignment = "RIGHT"
-        # layersRow.prop(gp.data, "layers")
-        layersRow.label(text="Layer:")
-        # layersRow.prop(prefs, "layersListDropdown", text="Layers")
-
-        # Grease pencil layer.
-        gpl = context.active_gpencil_layer
-        if gpl and gpl.info is not None:
-            text = gpl.info
-            maxw = 25
-            if len(text) > maxw:
-                text = text[: maxw - 5] + ".." + text[-3:]
-        else:
-            text = ""
-
-        # layout.label(text="Layer:")
-        sub = layersRow.row(align=True)
-        # sub.alignment = "LEFT"
-        # sub.scale_x = 0.8
-        sub.ui_units_x = 6
-        sub.popover(
-            panel="TOPBAR_PT_gpencil_layers",
-            text=text,
-        )
-        if gpl and gpl.info is not None:
-            sub.alert = gpl.lock
-            sub.prop(gpl, "lock", text="")
-            sub.alert = False
-        else:
-            sub.label(text="", icon="LOCKED")
-
-        gpOpsRow = layersRow.row(align=False)
-        gpOpsRow.separator(factor=0.1)
-        gpOpsLeftRow = gpOpsRow.row(align=True)
-        gpOpsLeftRow.ui_units_x = 3
-        gpOpsLeftRow.scale_x = 2.0
-        # gpOpsLeftRow.alignment = "RIGHT"
-        gpOpsLeftRow.operator("uas_shot_manager.select_grease_pencil_object", text="", icon="RESTRICT_SELECT_OFF")
-
-        if gp.mode == "PAINT_GPENCIL":
-            icon = "GREASEPENCIL"
-            gpOpsLeftRow.alert = True
-            # row.operator("uas_shot_manager.greasepencilitem", text="", icon=icon).index = index
-
-            gpOpsLeftRow.operator("uas_shot_manager.toggle_grease_pencil_draw_mode", text="", icon=icon)
-            # bpy.ops.gpencil.paintmode_toggle()
-        else:
-            icon = config.icons_col["ShotManager_CamGPShot_32"]
-            gpOpsLeftRow.operator("uas_shot_manager.toggle_grease_pencil_draw_mode", text="", icon_value=icon.icon_id)
-            # mainRow.operator("uas_shot_manager.draw_on_grease_pencil", text="", icon_value=icon.icon_id)
-            # row.operator("uas_shot_manager.greasepencilitem", text="", icon_value=icon.icon_id).index = index
-
-        gpOpsRightRow = gpOpsRow.row(align=False)
-        # gpOpsRightRow.separator(factor=0.1)
-        gpOpsRightRow.operator("uas_shot_manager.clear_layer", text="", icon="MESH_PLANE")
-
-    # Active material ###
-    ###################
-
-    materialsRow = box.row(align=True)
-    if objIsGP:
-        materialsRow.label(text="Material:")
-        # materialsRow.prop(gp, "material_slots")
-        materialsRow.prop(props, "greasePencil_activeMaterial", text="")
-
-    # current frame ###
-    ###################
-
-    mainRow = box.row(align=False)
-    keysRow = mainRow.row(align=True)
-    keysRow.scale_x = 1.2
-    keysRow.alignment = "CENTER"
-    keysRow.enabled = objIsGP
-    keysRow.operator("uas_shot_manager.greasepencil_previouskey", icon="PREV_KEYFRAME", text="")
-
-    isCurrentFrameOnGPFrame = False
-    if objIsGP:
-        isCurrentFrameOnGPFrame = utils_greasepencil.isCurrentFrameOnLayerFrame(
-            gp, context.scene.frame_current, props.greasePencil_layersMode
-        )
-    else:
-        keysRow.enabled = False
-
-    if isCurrentFrameOnGPFrame:
-        # iconFrame = "HANDLETYPE_FREE_VEC"
-        iconFrame = "ADD"
-        iconFrame = "KEYFRAME_HLT"
-    else:
-        iconFrame = "KEYTYPE_MOVING_HOLD_VEC"
-        iconFrame = "KEYFRAME"
-    # iconFrame = "ADD"
-    keysRow.operator("uas_shot_manager.greasepencil_addnewframe", icon=iconFrame, text="")
-    # keysRow.enabled = True
-    keysRow.operator("uas_shot_manager.greasepencil_nextkey", icon="NEXT_KEYFRAME", text="")
-
-    # subRow = mainRow.row(align=False)
-    # subRow.scale_x = 1.5
-    # subRow.alignment = "RIGHT"
-    # mainRow.scale_x = 2
-    subsubRow = mainRow.row(align=True)
-    # subsubRow.label(text="Apply to:")
-    subsubRow.scale_x = 0.9
-    # subRow.ui_units_x = 14
-    subsubRow.alignment = "LEFT"
-    subsubRow.prop(props, "greasePencil_layersMode", text="Apply to")
-
-    mainRow.label(text="Drawing on GP frame: ")
-    subsubRow = mainRow.row(align=True)
-    gpFrameStr = "-"
-    if objIsGP:
-        if isCurrentFrameOnGPFrame:
-            gpFrameStr = "Current"
-        else:
-            subsubRow.alert = True
-            gpFrameStr = str(utils_greasepencil.getLayerPreviousFrame(gp, context.scene.frame_current, "ACTIVE"))
-    subsubRow.label(text=gpFrameStr)
-
-    utils_ui.drawSeparatorLine(layout, lower_height=0.6)
-
-    # box.operator("uas_shot_manager.draw_on_grease_pencil", text="", icon="GP_SELECT_STROKES")